+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Compiler/FormPassTest.php
@@ -59,7 +59,10 @@ class FormPassTest extends \PHPUnit_Framework_TestCase
         ), $extDefinition->getArgument(1));
     }
 
-    public function testAddTaggedTypeExtensions()
+    /**
+     * @dataProvider addTaggedTypeExtensionsDataProvider
+     */
+    public function testAddTaggedTypeExtensions(array $extensions, array $expectedRegisteredExtensions)
     {
         $container = new ContainerBuilder();
         $container->addCompilerPass(new FormPass());
@@ -72,26 +75,49 @@ class FormPassTest extends \PHPUnit_Framework_TestCase
         ));
 
         $container->setDefinition('form.extension', $extDefinition);
-        $container->register('my.type_extension1', 'stdClass')
-            ->addTag('form.type_extension', array('extended_type' => 'type1'));
-        $container->register('my.type_extension2', 'stdClass')
-            ->addTag('form.type_extension', array('extended_type' => 'type1'));
-        $container->register('my.type_extension3', 'stdClass')
-            ->addTag('form.type_extension', array('extended_type' => 'type2'));
+
+        foreach ($extensions as $serviceId => $tag) {
+            $container->register($serviceId, 'stdClass')->addTag('form.type_extension', $tag);
+        }
 
         $container->compile();
 
         $extDefinition = $container->getDefinition('form.extension');
+        $this->assertSame($expectedRegisteredExtensions, $extDefinition->getArgument(2));
+    }
 
-        $this->assertSame(array(
-            'type1' => array(
-                'my.type_extension1',
-                'my.type_extension2',
+    /**
+     * @return array
+     */
+    public function addTaggedTypeExtensionsDataProvider()
+    {
+        return array(
+            array(
+                array(
+                    'my.type_extension1' => array('extended_type' => 'type1'),
+                    'my.type_extension2' => array('extended_type' => 'type1'),
+                    'my.type_extension3' => array('extended_type' => 'type2'),
+                ),
+                array(
+                    'type1' => array('my.type_extension1', 'my.type_extension2'),
+                    'type2' => array('my.type_extension3'),
+                ),
             ),
-            'type2' => array(
-                'my.type_extension3',
+            array(
+                array(
+                    'my.type_extension1' => array('extended_type' => 'type1', 'priority' => 1),
+                    'my.type_extension2' => array('extended_type' => 'type1', 'priority' => 2),
+                    'my.type_extension3' => array('extended_type' => 'type1', 'priority' => -1),
+                    'my.type_extension4' => array('extended_type' => 'type2', 'priority' => 2),
+                    'my.type_extension5' => array('extended_type' => 'type2', 'priority' => 1),
+                    'my.type_extension6' => array('extended_type' => 'type2', 'priority' => 1),
+                ),
+                array(
+                    'type1' => array('my.type_extension2', 'my.type_extension1', 'my.type_extension3'),
+                    'type2' => array('my.type_extension4', 'my.type_extension5', 'my.type_extension6'),
+                ),
             ),
-        ), $extDefinition->getArgument(2));
+        );
     }
 
     /**